refactor: HandlerScanner 분리
